Enable client authentication tests
Move those tests from a client-pool-tests.js file.
